+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.34.0 (August 08, 2019)
 
+* [ROCK-459] Fix error with Decimal types in Crossdata2 driver
 * [ROCK-475] Fix "Discovery goes down while testing stability"
 * [ROCK-475] Fix when Discovery showed only tables from default database.
 * [ROCK-148] Multitenant